extract _compilation_hash
