+++ src/gt4py/next/program_processors/runners/dace/workflow/backend.py
@@ -12,34 +12,13 @@ import factory
 
 import gt4py.next.allocators as next_allocators
 from gt4py._core import definitions as core_defs
-from gt4py.eve.utils import content_hash
 from gt4py.next import backend
-from gt4py.next.otf import stages, workflow
+from gt4py.next.otf import workflow
+from gt4py.next.program_processors.runners import compiled_backend
 from gt4py.next.program_processors.runners.dace.workflow.factory import DaCeWorkflowFactory
 
 
-def _compilation_hash(otf_closure: stages.CompilableProgram) -> int:
-    """Given closure compute a hash uniquely determining if we need to recompile.
-
-    This function was copied from 'next.program_processors.runners.gtfn'
-    """
-    offset_provider = otf_closure.args.offset_provider
-    return hash(
-        (
-            otf_closure.data,
-            # As the frontend types contain lists they are not hashable. As a workaround we just
-            # use content_hash here.
-            content_hash(tuple(arg for arg in otf_closure.args.args)),
-            # Directly using the `id` of the offset provider is not possible as the decorator adds
-            # the implicitly defined ones (i.e. to allow the `TDim + 1` syntax) resulting in a
-            # different `id` every time. Instead use the `id` of each individual offset provider.
-            tuple((k, id(v)) for (k, v) in offset_provider.items()) if offset_provider else None,
-            otf_closure.args.column_axis,
-        )
-    )
-
-
-class DaCeFieldviewBackendFactory(factory.Factory):
+class DaCeFieldviewBackendFactory(compiled_backend.CompiledBackendFactory):
     class Meta:
         model = backend.Backend
 
@@ -59,7 +38,6 @@ class DaCeFieldviewBackendFactory(factory.Factory):
             name_cached="_cached",
         )
         device_type = core_defs.DeviceType.CPU
-        hash_function = _compilation_hash
         otf_workflow = factory.SubFactory(
             DaCeWorkflowFactory,
             device_type=factory.SelfAttribute("..device_type"),
