+++ src/gt4py/next/program_processors/runners/gtfn.py
@@ -19,7 +19,6 @@ import filelock
 import gt4py._core.definitions as core_defs
 import gt4py.next.allocators as next_allocators
 from gt4py.eve import utils
-from gt4py.eve.utils import content_hash
 from gt4py.next import backend, common, config
 from gt4py.next.iterator import ir as itir
 from gt4py.next.otf import arguments, recipes, stages, workflow
@@ -27,6 +26,7 @@ from gt4py.next.otf.binding import nanobind
 from gt4py.next.otf.compilation import compiler
 from gt4py.next.otf.compilation.build_systems import compiledb
 from gt4py.next.program_processors.codegens.gtfn import gtfn_module
+from gt4py.next.program_processors.runners import compiled_backend
 
 
 # TODO(ricoh): Add support for the whole range of arguments that can be passed to a fencil.
@@ -100,24 +100,6 @@ def extract_connectivity_args(
                 f"but got '{type(conn).__name__}'."
             )
     return args
-
-
-def compilation_hash(otf_closure: stages.CompilableProgram) -> int:
-    """Given closure compute a hash uniquely determining if we need to recompile."""
-    offset_provider = otf_closure.args.offset_provider
-    return hash(
-        (
-            otf_closure.data,
-            # As the frontend types contain lists they are not hashable. As a workaround we just
-            # use content_hash here.
-            content_hash(tuple(arg for arg in otf_closure.args.args)),
-            # Directly using the `id` of the offset provider is not possible as the decorator adds
-            # the implicitly defined ones (i.e. to allow the `TDim + 1` syntax) resulting in a
-            # different `id` every time. Instead use the `id` of each individual offset provider.
-            tuple((k, id(v)) for (k, v) in offset_provider.items()) if offset_provider else None,
-            otf_closure.args.column_axis,
-        )
-    )
 
 
 def fingerprint_compilable_program(inp: stages.CompilableProgram) -> str:
@@ -215,7 +197,7 @@ class GTFNCompileWorkflowFactory(factory.Factory):
     )
 
 
-class GTFNBackendFactory(factory.Factory):
+class GTFNBackendFactory(compiled_backend.CompiledBackendFactory):
     class Meta:
         model = backend.Backend
 
@@ -236,7 +218,6 @@ class GTFNBackendFactory(factory.Factory):
             name_cached="_cached",
         )
         device_type = core_defs.DeviceType.CPU
-        hash_function = compilation_hash
         otf_workflow = factory.SubFactory(
             GTFNCompileWorkflowFactory, device_type=factory.SelfAttribute("..device_type")
         )
